# GH-38542: [C++][Parquet] Faster scalar BYTE_STREAM_SPLIT