@pitrou 2023-10-31
@cyb70289 Do you want to take a look at how this performs on ARM?