@pitrou 2023-10-31
Note that one reason for this work is that we might want to generalize BYTE_STREAM_SPLIT for more datatypes.