@cyb70289 2023-11-01
Big performance improvement observed on Arm Neoverse-N2 CPU. Great work!

- before
```
BM_ByteStreamSplitDecode_Float_Scalar/1024         2241 ns         2241 ns       312363 bytes_per_second=1.70208G/s
BM_ByteStreamSplitDecode_Float_Scalar/4096         8949 ns         8949 ns        78233 bytes_per_second=1.70515G/s
BM_ByteStreamSplitDecode_Float_Scalar/32768       71555 ns        71557 ns         9780 bytes_per_second=1.70592G/s
BM_ByteStreamSplitDecode_Float_Scalar/65536      143178 ns       143174 ns         4889 bytes_per_second=1.7052G/s

BM_ByteStreamSplitDecode_Double_Scalar/1024        4035 ns         4035 ns       173607 bytes_per_second=1.89097G/s
BM_ByteStreamSplitDecode_Double_Scalar/4096       16115 ns        16114 ns        43351 bytes_per_second=1.89381G/s
BM_ByteStreamSplitDecode_Double_Scalar/32768     138440 ns       138442 ns         5055 bytes_per_second=1.76348G/s
BM_ByteStreamSplitDecode_Double_Scalar/65536     277661 ns       277651 ns         2522 bytes_per_second=1.75862G/s

BM_ByteStreamSplitEncode_Float_Scalar/1024         2275 ns         2275 ns       307740 bytes_per_second=1.6771G/s
BM_ByteStreamSplitEncode_Float_Scalar/4096         9107 ns         9107 ns        77037 bytes_per_second=1.67555G/s
BM_ByteStreamSplitEncode_Float_Scalar/32768       72365 ns        72364 ns         9670 bytes_per_second=1.68689G/s
BM_ByteStreamSplitEncode_Float_Scalar/65536      145514 ns       145518 ns         4809 bytes_per_second=1.67773G/s

BM_ByteStreamSplitEncode_Double_Scalar/1024        6262 ns         6262 ns       111792 bytes_per_second=1.21838G/s
BM_ByteStreamSplitEncode_Double_Scalar/4096       25252 ns        25253 ns        27710 bytes_per_second=1.20847G/s
BM_ByteStreamSplitEncode_Double_Scalar/32768     162099 ns       162096 ns         4319 bytes_per_second=1.50614G/s
BM_ByteStreamSplitEncode_Double_Scalar/65536     323703 ns       323682 ns         2162 bytes_per_second=1.50852G/s
```

- after
```
BM_ByteStreamSplitDecode_Float_Scalar/1024         1129 ns         1129 ns       619351 bytes_per_second=3.37958G/s
BM_ByteStreamSplitDecode_Float_Scalar/4096         4499 ns         4499 ns       155591 bytes_per_second=3.39128G/s
BM_ByteStreamSplitDecode_Float_Scalar/32768       36209 ns        36210 ns        19331 bytes_per_second=3.37113G/s
BM_ByteStreamSplitDecode_Float_Scalar/65536       72520 ns        72521 ns         9650 bytes_per_second=3.36649G/s

BM_ByteStreamSplitDecode_Double_Scalar/1024        2999 ns         2999 ns       233336 bytes_per_second=2.54407G/s
BM_ByteStreamSplitDecode_Double_Scalar/4096       11979 ns        11979 ns        58390 bytes_per_second=2.54754G/s
BM_ByteStreamSplitDecode_Double_Scalar/32768      96324 ns        96323 ns         7266 bytes_per_second=2.5346G/s
BM_ByteStreamSplitDecode_Double_Scalar/65536     194568 ns       194574 ns         3595 bytes_per_second=2.50949G/s

BM_ByteStreamSplitEncode_Float_Scalar/1024          936 ns          936 ns       744854 bytes_per_second=4.07715G/s
BM_ByteStreamSplitEncode_Float_Scalar/4096         3705 ns         3705 ns       188993 bytes_per_second=4.11788G/s
BM_ByteStreamSplitEncode_Float_Scalar/32768       29817 ns        29816 ns        23459 bytes_per_second=4.09408G/s
BM_ByteStreamSplitEncode_Float_Scalar/65536       60051 ns        60050 ns        11632 bytes_per_second=4.06565G/s

BM_ByteStreamSplitEncode_Double_Scalar/1024        1872 ns         1871 ns       373845 bytes_per_second=4.07669G/s
BM_ByteStreamSplitEncode_Double_Scalar/4096        7471 ns         7471 ns        93678 bytes_per_second=4.0849G/s
BM_ByteStreamSplitEncode_Double_Scalar/32768      60360 ns        60362 ns        11600 bytes_per_second=4.04463G/s
BM_ByteStreamSplitEncode_Double_Scalar/65536     121104 ns       121107 ns         5776 bytes_per_second=4.0318G/s
```